+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/prompt/PromptRequest.kt
@@ -82,4 +82,44 @@ sealed class PromptRequest {
         val onMultipleFilesSelected: (Context, Array<Uri>) -> Unit,
         val onDismiss: () -> Unit
     ) : PromptRequest()
+
+    /**
+     * Value type that represents a request for an authentication prompt.
+     * For more related info take a look at
+     * <a href="https://developer.mozilla.org/en-US/docs/Web/HTTP/Authentication>MDN docs</a>
+     * @property title of the dialog.
+     * @property message the body of the dialog.
+     * @property userName default value provide for this session.
+     * @property password default value provide for this session.
+     * @property method type of authentication,  valid values [Method.HOST] and [Method.PROXY].
+     * @property level indicates the level of security of the authentication like [Level.NONE],
+     * [Level.SECURED] and [Level.PASSWORD_ENCRYPTED].
+     * @property onlyShowPassword indicates if the dialog should only include a password field.
+     * @property previousFailed indicates if this request is the result of a previous failed attempt to login.
+     * @property isCrossOrigin indicates if this request is from a cross-origin sub-resource.
+     * @property onConfirm callback to indicate the user want to start the authentication flow.
+     * @property onDismiss callback to indicate the user dismissed this request.
+     */
+    data class Authentication(
+        val title: String,
+        val message: String,
+        val userName: String,
+        val password: String,
+        val method: Method,
+        val level: Level,
+        val onlyShowPassword: Boolean = false,
+        val previousFailed: Boolean = false,
+        val isCrossOrigin: Boolean = false,
+        val onConfirm: (String, String) -> Unit,
+        val onDismiss: () -> Unit
+    ) : PromptRequest() {
+
+        enum class Level {
+            NONE, PASSWORD_ENCRYPTED, SECURED
+        }
+
+        enum class Method {
+            HOST, PROXY
+        }
+    }
 }